feat: 영화 고유 값을 string 타입으로 변경

+++ Architecture_Pattern_Study/app/src/main/java/com/joohyeong/architecture_pattern_study/MainActivity.kt
@@ -38,7 +38,7 @@ class MainActivity : AppCompatActivity() {
         binding.recyclerMovie.adapter = MovieAdapter(
             movies = movies,
             onMovieClick = {
-                startActivity(MovieDetailActivity.newIntent(context = this, movieId = it.toLong()))
+                startActivity(MovieDetailActivity.newIntent(context = this, movieId = it))
             }
         )
     }
